# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Karat Hive API",
    "version": "0.1.0",
    "description": "Generated from Nest controllers + ZodValidationPipe schemas (`npm run openapi:generate`). Zod remains the single source of request shape (NFR-030, AD-BE-14). First artefact is partial: paths/methods and request Zod schemas are covered; response presenters are not yet dual-exported as Zod. Pre-code catalogue: docs/API-Route-Inventory.md — clients should treat this file as the evolving contract."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Karat Hive access JWT (not a Firebase ID token)."
      }
    },
    "schemas": {
      "SuccessMeta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "serverTime": {
            "type": "string"
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "requestId",
          "serverTime",
          "nextCursor"
        ]
      },
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/SuccessMeta"
          }
        },
        "required": [
          "meta"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "nullable": true
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    },
...
  },
  "paths": {
...
    "/v1/admin/connections": {
      "get": {
        "operationId": "Admin_listConnections",
        "tags": [
          "Admin"
        ],
        "summary": "GET /v1/admin/connections",
        "description": "Handler Admin.listConnections. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/connections/{id}": {
      "get": {
        "operationId": "Admin_getConnection",
        "tags": [
          "Admin"
        ],
        "summary": "GET /v1/admin/connections/{id}",
        "description": "Handler Admin.getConnection. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/connections/{id}": {
      "get": {
        "operationId": "Connection_getConnectionById",
        "tags": [
          "Connection"
        ],
        "summary": "GET /v1/connections/{id}",
        "description": "Handler Connection.getConnectionById. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connections/{id}/close": {
      "post": {
        "operationId": "Connection_closeConnection",
        "tags": [
          "Connection"
        ],
        "summary": "POST /v1/connections/{id}/close",
        "description": "Handler Connection.closeConnection. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connections/{id}/contact-events": {
      "post": {
        "operationId": "Connection_recordContactEvent",
        "tags": [
          "Connection"
        ],
        "summary": "POST /v1/connections/{id}/contact-events",
        "description": "Handler Connection.recordContactEvent. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "channel": {
                    "type": "string",
                    "enum": [
                      "WHATSAPP",
                      "PHONE"
                    ]
                  }
                },
                "required": [
                  "channel"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/devices": {
      "post": {
        "operationId": "Device_registerDevice",
        "tags": [
          "Device"
        ],
        "summary": "POST /v1/devices",
        "description": "Handler Device.registerDevice. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "platform": {
                    "type": "string",
                    "enum": [
                      "IOS",
                      "ANDROID"
                    ]
                  },
                  "pushToken": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "label": {
                    "type": "string",
                    "maxLength": 100
                  }
                },
                "required": [
                  "platform",
                  "pushToken"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/devices/{id}": {
      "delete": {
        "operationId": "Device_deleteDevice",
        "tags": [
          "Device"
        ],
        "summary": "DELETE /v1/devices/{id}",
        "description": "Handler Device.deleteDevice. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/me/connections": {
      "get": {
        "operationId": "Connection_listMyConnections",
        "tags": [
          "Connection"
        ],
        "summary": "GET /v1/me/connections",
        "description": "Handler Connection.listMyConnections. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "CLOSED"
              ]
            },
            "required": false,
            "name": "state",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/offers/{id}/accept": {
      "post": {
        "operationId": "Connection_acceptOffer",
        "tags": [
          "Connection"
        ],
        "summary": "POST /v1/offers/{id}/accept",
        "description": "Handler Connection.acceptOffer. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "confirmation": {
                    "type": "string",
                    "enum": [
                      "REVEAL_AND_CONNECT"
                    ]
                  }
                },
                "required": [
                  "confirmation"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  }
}